chore: moved ConfigKey.kt to new config directory
