+++ app/src/main/kotlin/org/coremapx/app/userDirectory/ConfigRepository.kt
@@ -10,28 +10,28 @@ import mu.KotlinLogging
 import org.coremapx.app.config.PrivateConfig
 import org.coremapx.app.theme.DefaultThemes
 import org.coremapx.app.theme.ThemeConfig
-import org.coremapx.app.userDirectory.ConfigKeys.BACKGROUND
-import org.coremapx.app.userDirectory.ConfigKeys.BORDER_COLOR
-import org.coremapx.app.userDirectory.ConfigKeys.CANVAS_BACKGROUND_COLOR
-import org.coremapx.app.userDirectory.ConfigKeys.COMMAND_LINE_BACKGROUND_COLOR
-import org.coremapx.app.userDirectory.ConfigKeys.EDGE_MAIN_COLOR
-import org.coremapx.app.userDirectory.ConfigKeys.ERROR
-import org.coremapx.app.userDirectory.ConfigKeys.HOVERED_BORDER_COLOR
-import org.coremapx.app.userDirectory.ConfigKeys.ON_BACKGROUND
-import org.coremapx.app.userDirectory.ConfigKeys.ON_ERROR
-import org.coremapx.app.userDirectory.ConfigKeys.ON_PRIMARY
-import org.coremapx.app.userDirectory.ConfigKeys.ON_SECONDARY
-import org.coremapx.app.userDirectory.ConfigKeys.ON_SURFACE
-import org.coremapx.app.userDirectory.ConfigKeys.PRIMARY
-import org.coremapx.app.userDirectory.ConfigKeys.PRIMARY_VARIANT
-import org.coremapx.app.userDirectory.ConfigKeys.SECONDARY
-import org.coremapx.app.userDirectory.ConfigKeys.SECONDARY_VARIANT
-import org.coremapx.app.userDirectory.ConfigKeys.SUCCESS_COLOR
-import org.coremapx.app.userDirectory.ConfigKeys.SURFACE
-import org.coremapx.app.userDirectory.ConfigKeys.SYSTEM_DIALOG_THEME
-import org.coremapx.app.userDirectory.ConfigKeys.THEME
-import org.coremapx.app.userDirectory.ConfigKeys.VERTEX_MAIN_COLOR
-import org.coremapx.app.userDirectory.ConfigKeys.WARNING_COLOR
+import org.coremapx.app.userDirectory.config.ConfigKeys.BACKGROUND
+import org.coremapx.app.userDirectory.config.ConfigKeys.BORDER_COLOR
+import org.coremapx.app.userDirectory.config.ConfigKeys.CANVAS_BACKGROUND_COLOR
+import org.coremapx.app.userDirectory.config.ConfigKeys.COMMAND_LINE_BACKGROUND_COLOR
+import org.coremapx.app.userDirectory.config.ConfigKeys.EDGE_MAIN_COLOR
+import org.coremapx.app.userDirectory.config.ConfigKeys.ERROR
+import org.coremapx.app.userDirectory.config.ConfigKeys.HOVERED_BORDER_COLOR
+import org.coremapx.app.userDirectory.config.ConfigKeys.ON_BACKGROUND
+import org.coremapx.app.userDirectory.config.ConfigKeys.ON_ERROR
+import org.coremapx.app.userDirectory.config.ConfigKeys.ON_PRIMARY
+import org.coremapx.app.userDirectory.config.ConfigKeys.ON_SECONDARY
+import org.coremapx.app.userDirectory.config.ConfigKeys.ON_SURFACE
+import org.coremapx.app.userDirectory.config.ConfigKeys.PRIMARY
+import org.coremapx.app.userDirectory.config.ConfigKeys.PRIMARY_VARIANT
+import org.coremapx.app.userDirectory.config.ConfigKeys.SECONDARY
+import org.coremapx.app.userDirectory.config.ConfigKeys.SECONDARY_VARIANT
+import org.coremapx.app.userDirectory.config.ConfigKeys.SUCCESS_COLOR
+import org.coremapx.app.userDirectory.config.ConfigKeys.SURFACE
+import org.coremapx.app.userDirectory.config.ConfigKeys.SYSTEM_DIALOG_THEME
+import org.coremapx.app.userDirectory.config.ConfigKeys.THEME
+import org.coremapx.app.userDirectory.config.ConfigKeys.VERTEX_MAIN_COLOR
+import org.coremapx.app.userDirectory.config.ConfigKeys.WARNING_COLOR
 import java.io.File
 import java.util.Properties
