+++ app/src/main/kotlin/org/coremapx/app/userDirectory/ConfigValidator.kt
@@ -3,50 +3,50 @@ package org.coremapx.app.userDirectory
 import extensions.toColorOrNull
 import model.result.ConfigErrors
 import model.result.Result
-import org.coremapx.app.userDirectory.ConfigKeys.ANIMATION_DURATION
-import org.coremapx.app.userDirectory.ConfigKeys.BACKGROUND
-import org.coremapx.app.userDirectory.ConfigKeys.BORDER_COLOR
-import org.coremapx.app.userDirectory.ConfigKeys.CANVAS_BACKGROUND_COLOR
-import org.coremapx.app.userDirectory.ConfigKeys.CANVAS_DRAG_RATIO
-import org.coremapx.app.userDirectory.ConfigKeys.CANVAS_LIMIT
-import org.coremapx.app.userDirectory.ConfigKeys.COMMAND_FIELD_SCROLL_DELAY
-import org.coremapx.app.userDirectory.ConfigKeys.COMMAND_FIELD_WIDTH
-import org.coremapx.app.userDirectory.ConfigKeys.COMMAND_LINE_BACKGROUND_COLOR
-import org.coremapx.app.userDirectory.ConfigKeys.EDGE_ARROW_SIZE
-import org.coremapx.app.userDirectory.ConfigKeys.EDGE_LABEL_SIZE
-import org.coremapx.app.userDirectory.ConfigKeys.EDGE_MAIN_COLOR
-import org.coremapx.app.userDirectory.ConfigKeys.EDGE_WIDTH
-import org.coremapx.app.userDirectory.ConfigKeys.ERROR
-import org.coremapx.app.userDirectory.ConfigKeys.GRAPH_LAYOUT_HEIGHT
-import org.coremapx.app.userDirectory.ConfigKeys.GRAPH_LAYOUT_WIDTH
-import org.coremapx.app.userDirectory.ConfigKeys.HOVERED_BORDER_COLOR
-import org.coremapx.app.userDirectory.ConfigKeys.IS_EXPANDED_SETTINGS
-import org.coremapx.app.userDirectory.ConfigKeys.IS_TRANSPARENT_COMMAND_LINE
-import org.coremapx.app.userDirectory.ConfigKeys.LANGUAGE
-import org.coremapx.app.userDirectory.ConfigKeys.MAIN_SCREEN_START_HEIGHT
-import org.coremapx.app.userDirectory.ConfigKeys.MAIN_SCREEN_START_WIDTH
-import org.coremapx.app.userDirectory.ConfigKeys.MAX_COUNT_MESSAGES
-import org.coremapx.app.userDirectory.ConfigKeys.MESSAGE_OUTPUT_HEIGHT
-import org.coremapx.app.userDirectory.ConfigKeys.ON_BACKGROUND
-import org.coremapx.app.userDirectory.ConfigKeys.ON_ERROR
-import org.coremapx.app.userDirectory.ConfigKeys.ON_PRIMARY
-import org.coremapx.app.userDirectory.ConfigKeys.ON_SECONDARY
-import org.coremapx.app.userDirectory.ConfigKeys.ON_SURFACE
-import org.coremapx.app.userDirectory.ConfigKeys.PRIMARY
-import org.coremapx.app.userDirectory.ConfigKeys.PRIMARY_VARIANT
-import org.coremapx.app.userDirectory.ConfigKeys.SECONDARY
-import org.coremapx.app.userDirectory.ConfigKeys.SECONDARY_VARIANT
-import org.coremapx.app.userDirectory.ConfigKeys.START_WINDOW_PLACEMENT
-import org.coremapx.app.userDirectory.ConfigKeys.SUCCESS_COLOR
-import org.coremapx.app.userDirectory.ConfigKeys.SURFACE
-import org.coremapx.app.userDirectory.ConfigKeys.SYSTEM_DIALOG_THEME
-import org.coremapx.app.userDirectory.ConfigKeys.THEME
-import org.coremapx.app.userDirectory.ConfigKeys.TITLE_BAR_HEIGHT
-import org.coremapx.app.userDirectory.ConfigKeys.TITLE_BAR_ICON_SIZE
-import org.coremapx.app.userDirectory.ConfigKeys.VERTEX_LABEL_SIZE
-import org.coremapx.app.userDirectory.ConfigKeys.VERTEX_MAIN_COLOR
-import org.coremapx.app.userDirectory.ConfigKeys.VERTEX_RADIUS
-import org.coremapx.app.userDirectory.ConfigKeys.WARNING_COLOR
+import org.coremapx.app.userDirectory.config.ConfigKeys.ANIMATION_DURATION
+import org.coremapx.app.userDirectory.config.ConfigKeys.BACKGROUND
+import org.coremapx.app.userDirectory.config.ConfigKeys.BORDER_COLOR
+import org.coremapx.app.userDirectory.config.ConfigKeys.CANVAS_BACKGROUND_COLOR
+import org.coremapx.app.userDirectory.config.ConfigKeys.CANVAS_DRAG_RATIO
+import org.coremapx.app.userDirectory.config.ConfigKeys.CANVAS_LIMIT
+import org.coremapx.app.userDirectory.config.ConfigKeys.COMMAND_FIELD_SCROLL_DELAY
+import org.coremapx.app.userDirectory.config.ConfigKeys.COMMAND_FIELD_WIDTH
+import org.coremapx.app.userDirectory.config.ConfigKeys.COMMAND_LINE_BACKGROUND_COLOR
+import org.coremapx.app.userDirectory.config.ConfigKeys.EDGE_ARROW_SIZE
+import org.coremapx.app.userDirectory.config.ConfigKeys.EDGE_LABEL_SIZE
+import org.coremapx.app.userDirectory.config.ConfigKeys.EDGE_MAIN_COLOR
+import org.coremapx.app.userDirectory.config.ConfigKeys.EDGE_WIDTH
+import org.coremapx.app.userDirectory.config.ConfigKeys.ERROR
+import org.coremapx.app.userDirectory.config.ConfigKeys.GRAPH_LAYOUT_HEIGHT
+import org.coremapx.app.userDirectory.config.ConfigKeys.GRAPH_LAYOUT_WIDTH
+import org.coremapx.app.userDirectory.config.ConfigKeys.HOVERED_BORDER_COLOR
+import org.coremapx.app.userDirectory.config.ConfigKeys.IS_EXPANDED_SETTINGS
+import org.coremapx.app.userDirectory.config.ConfigKeys.IS_TRANSPARENT_COMMAND_LINE
+import org.coremapx.app.userDirectory.config.ConfigKeys.LANGUAGE
+import org.coremapx.app.userDirectory.config.ConfigKeys.MAIN_SCREEN_START_HEIGHT
+import org.coremapx.app.userDirectory.config.ConfigKeys.MAIN_SCREEN_START_WIDTH
+import org.coremapx.app.userDirectory.config.ConfigKeys.MAX_COUNT_MESSAGES
+import org.coremapx.app.userDirectory.config.ConfigKeys.MESSAGE_OUTPUT_HEIGHT
+import org.coremapx.app.userDirectory.config.ConfigKeys.ON_BACKGROUND
+import org.coremapx.app.userDirectory.config.ConfigKeys.ON_ERROR
+import org.coremapx.app.userDirectory.config.ConfigKeys.ON_PRIMARY
+import org.coremapx.app.userDirectory.config.ConfigKeys.ON_SECONDARY
+import org.coremapx.app.userDirectory.config.ConfigKeys.ON_SURFACE
+import org.coremapx.app.userDirectory.config.ConfigKeys.PRIMARY
+import org.coremapx.app.userDirectory.config.ConfigKeys.PRIMARY_VARIANT
+import org.coremapx.app.userDirectory.config.ConfigKeys.SECONDARY
+import org.coremapx.app.userDirectory.config.ConfigKeys.SECONDARY_VARIANT
+import org.coremapx.app.userDirectory.config.ConfigKeys.START_WINDOW_PLACEMENT
+import org.coremapx.app.userDirectory.config.ConfigKeys.SUCCESS_COLOR
+import org.coremapx.app.userDirectory.config.ConfigKeys.SURFACE
+import org.coremapx.app.userDirectory.config.ConfigKeys.SYSTEM_DIALOG_THEME
+import org.coremapx.app.userDirectory.config.ConfigKeys.THEME
+import org.coremapx.app.userDirectory.config.ConfigKeys.TITLE_BAR_HEIGHT
+import org.coremapx.app.userDirectory.config.ConfigKeys.TITLE_BAR_ICON_SIZE
+import org.coremapx.app.userDirectory.config.ConfigKeys.VERTEX_LABEL_SIZE
+import org.coremapx.app.userDirectory.config.ConfigKeys.VERTEX_MAIN_COLOR
+import org.coremapx.app.userDirectory.config.ConfigKeys.VERTEX_RADIUS
+import org.coremapx.app.userDirectory.config.ConfigKeys.WARNING_COLOR
 
 class ConfigValidator {
     companion object {
