+++ app/src/main/kotlin/view/appInterface/settingsElements/blocks/TitleBarBlock.kt
@@ -12,8 +12,8 @@ import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.dp
 import org.coremapx.app.config
 import org.coremapx.app.theme.AppTheme
-import org.coremapx.app.userDirectory.ConfigKeys.TITLE_BAR_HEIGHT
-import org.coremapx.app.userDirectory.ConfigKeys.TITLE_BAR_ICON_SIZE
+import org.coremapx.app.userDirectory.config.ConfigKeys.TITLE_BAR_HEIGHT
+import org.coremapx.app.userDirectory.config.ConfigKeys.TITLE_BAR_ICON_SIZE
 import view.appInterface.preview.PreviewSurface
 import view.appInterface.settingsElements.lines.NumberTextFieldLine
